+++ packages_test.go
@@ -27,13 +27,13 @@ func TestStdlibPackages(t *testing.T) {
 				Name:      "archive/tar",
 				Constants: []string{"FormatGNU Format", "FormatPAX Format", "FormatUSTAR Format", "FormatUnknown Format", "TypeBlock untyped rune", "TypeChar untyped rune", "TypeCont untyped rune", "TypeDir untyped rune", "TypeFifo untyped rune", "TypeGNULongLink untyped rune", "TypeGNULongName untyped rune", "TypeGNUSparse untyped rune", "TypeLink untyped rune", "TypeReg untyped rune", "TypeRegA untyped rune", "TypeSymlink untyped rune", "TypeXGlobalHeader untyped rune", "TypeXHeader untyped rune"},
 				Variables: []string{"ErrFieldTooLong error", "ErrHeader error", "ErrWriteAfterClose error", "ErrWriteTooLong error"},
-				Functions: []string{"FileInfoHeader(fi os.FileInfo, link string) (*Header, error)", "NewReader(r io.Reader) *Reader", "NewWriter(w io.Writer) *Writer"},
+				Functions: []string{"FileInfoHeader(fi io/fs.FileInfo, link string) (*Header, error)", "NewReader(r io.Reader) *Reader", "NewWriter(w io.Writer) *Writer"},
 				Types: map[string][]string{
 					"Format int": {
 						"(Format) String() string",
 					},
 					"Header struct": {
-						"(*Header) FileInfo() os.FileInfo",
+						"(*Header) FileInfo() io/fs.FileInfo",
 					},
 					"Reader struct": {
 						"(*Reader) Next() (*Header, error)",
@@ -51,13 +51,13 @@ func TestStdlibPackages(t *testing.T) {
 	}
 
 	for _, v := range tt {
-		pak, err := newPak(v.importPath)
+		p, err := newPak(v.importPath)
 		if err != nil {
 			t.Errorf("\ngot \n\t%#+v \nwanted \n\t%#+v", err, v.expected)
 		}
 
-		if !cmp.Equal(pak, v.expected) {
-			t.Errorf("\ngot \n\t%#+v \nwanted \n\t%#+v", pak, v.expected)
+		if !cmp.Equal(p, v.expected) {
+			t.Error(cmp.Diff(v.expected, p))
 		}
 	}
 }
@@ -105,7 +105,7 @@ func TestThirdPartyPackages(t *testing.T) {
 		}
 
 		if !cmp.Equal(pak, v.expected) {
-			t.Errorf("\ngot \n\t%#+v \nwanted \n\t%#+v", pak, v.expected)
+			t.Error(cmp.Diff(v.expected, pak))
 		}
 	}
 }