+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Most recent version is listed first.
 
 ## v0.0.2
 - add test example: https://github.com/komuw/kama/pull/13
+- add types to the fields of a struct: https://github.com/komuw/kama/pull/16
 
 ## v0.0.1
 - pretty print variables and packages: https://github.com/komuw/kama/pull/10